+++ src/max_div/_core/metrics/_distance/__init__.py
@@ -7,4 +7,5 @@ from ._store import (
     DISTANCE_STORE_TYPE,
     DistanceStore,
     get_distance,
+    lazy_store,
 )